Fix cache usage #36, #23
