+++ src/main/kotlin/com/xcart/prognosis/repositories/XbApi.kt
@@ -1,0 +1,50 @@
+package com.xcart.prognosis.repositories
+
+import com.fasterxml.jackson.databind.DeserializationFeature
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.module.kotlin.registerKotlinModule
+import com.github.kittinunf.fuel.Fuel
+import com.github.kittinunf.result.Result
+import com.xcart.prognosis.domain.VacationInfoRef
+import com.xcart.prognosis.domain.VacationPeriod
+import com.xcart.prognosis.errors.ExternalServiceError
+import com.xcart.prognosis.services.Configuration
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.cache.annotation.CacheEvict
+import org.springframework.cache.annotation.Cacheable
+import org.springframework.scheduling.annotation.Scheduled
+import org.springframework.stereotype.Service
+
+@Service
+class XbApi @Autowired constructor(config: Configuration) {
+    private val baseUrl = config.xbUrl
+    private val xbPassword = config.xbPassword
+
+    @Cacheable("getAllVacations")
+    fun getVacationsInfo(): List<VacationPeriod> {
+        val params = listOf(
+                "target" to "vacations_info",
+                "password" to xbPassword
+        )
+        try {
+            val (_, _, result) = Fuel.get(baseUrl, params).responseString()
+            return when (result) {
+                is Result.Failure -> throw result.getException()
+                else -> deserializeVacations(result.get())
+            }
+        } catch (ex: Exception) {
+            throw ExternalServiceError("Error during communication with XB API", ex)
+        }
+    }
+
+    @CacheEvict("vacationsInfo")
+    @Scheduled(fixedDelay = 3600000)
+    fun cacheEvict() {}
+
+    fun deserializeVacations(json: String): List<VacationPeriod> {
+        val mapper = ObjectMapper().registerKotlinModule()
+                .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+        val result = mapper.readValue(json, VacationInfoRef())
+        return result.map { it.value }
+    }
+}
