+++ src/main/kotlin/com/xcart/prognosis/services/CacheConfiguration.kt
@@ -1,0 +1,9 @@
+package com.xcart.prognosis.services
+
+import org.springframework.cache.annotation.EnableCaching
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@EnableCaching
+class CacheConfiguration {
+}
