+++ src/test/kotlin/com/xcart/prognosis/logic/WorkloadAnalysisTest.kt
@@ -7,9 +7,11 @@ import com.xcart.prognosis.domain.User
 import com.xcart.prognosis.reports.workload.DailyWorkloadItem
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
 import java.sql.Timestamp
 import java.time.LocalDate
 
+@SpringBootTest
 internal class WorkloadAnalysisTest {
 
     @Test
